PICARD: add and use OUTPUT_NAME recipe parameter
If a single output file will be generated by the current invocation of
PICARD, the OUTPUT_NAME recipe parameter will be used (if given) to
determine the name of that output file. If not specified, or if
multiple files are being processed, the suffix will be appended to the
current Frame file (rather than replacing an existing suffix).