SCUBA2: pass existing matched-filtered map to peak finder
As part of the FCF calculation, this primitive applies the matched
filter to the input map. This map can be passed on to _FIND_MAP_PEAK_
for its initial estimate of the source position, reducing the number
of applications of the matched filter by 2 and providing speed gains
of 1-2 seconds in the QL pipeline.