SCUBA2: proceed with assessment if a peak has been found
For calibrators, the process of calculating FCFs sets a Frame uhdr
entry called MAP_PEAK_POSITION which stores the fitted peak if such a
fit was successful. Given that a peak has been found if this flag is
set, this can be used as a way of proceeding with the assessment
beyond the detection of the source at a set s/n (which is estimated
from the un-matched-filtered map and is thus prone to not detecting
low s/n sources which show up clearly in the matched-filtered
version).

This should eliminate all those warning messages that the pipeline
could not locate a point source for calculating pointing offsets
(which is clearly ridiculous if FCFs have been successfully
calculated).