SCUBA2: rewrite to better deal with summit pipeline
This primitive was originally written to deal with offline analysis of
the two fastflat ramps associated with each observation. Only later
was it pressed into service as doing the same for the online
pipelines. As such the code was incapable of properly dealing with a
situation that often arises in the summit pipeline where the input
Frame object includes fastflat and science data (and sometimes
multiple fastflats). The primitive now expects this as the default and
will separate out the science data for later processing, dealing with
all the fastflat data here. Each fastflat ramp is processed and
analyzed separately, with the results (and raw flats) stored in the
calibration system.

Copies of the raw fastflats are left on disk, along with the solutions
for later downstream processing.