SCUBA2: store flatfield solutions in uhdr
In the case where the pipeline is processing a mix of science and
fastflat data, it cannot use the calibration system to retrieve the
fastflats. In this case, store the filenames in a hash indexed by
subarray and store that in the Frame uhdr under FLATS which other
primitives will then look for.