SCUBA2: issue supplementary error message in QL/SUMMIT pipelines
The error which results from the case where flatfield solutions were
found but no raw fastflats can be triggered in the QL and SUMMIT
pipelines if the fastflat at the beginning of the observation was
skipped. Therefore, print a supplemental error message indicating the
possible cause of this error for the QL and SUMMIT pipelines.