PICARD: use the KEEPFILES parameter to leave PSF on disk
If the recipe is creating its own PSF to use then the user may wish to
keep it on disk for analysis or re-use. The existing recipe paramter
KEEPFILES can be used to do this. A value of 1 indicates the PSF file
will remain on disk - any other value and it will be deleted.