SCUBA2: only set PROCESS_FASTFLAT to 1 for single-sequence data
The PROCESS_FASTFLAT flag is only used by other primitives when the
Frame contains single-sequence data to control whether or not those
steps are taken. Mixed-sequence data is already dealt with here and
subsequent processing will need to go ahead for the other (presumably
science) data, so there is no need to set PROCESS_FASTFLAT to true.