SCUBA2: remove creation of bad-bolometer masks
The map-maker does a perfectly good job of weeding out the worst
bolometers so the creation of bad bolometer masks is not necessary,
other than for analysis and there is no evidence that anyone is
looking at these files. The responsivity map is effectively a
bad-bolometer mask, though one with all bolometers rather than a
subset of the best.