Optimized external index handling for StripCharts.
Added a :SINCE filter to ORAC::Index::scanindex which restricts the
output to entries with ORACTIME greater than this value.

Allowed indexentry to take the index as a hash reference argument.
This avoids it having to call indexref, which, in the case of
ORAC::Index::External, causes the status of the file to be checked.

ORAC::Index::Extern::indexref now only checks the file if it was not
given a hash reference, because when it does check the file, it calls
slurpindex which calls it again with the data.

Added caches RulesCount and RulesKeys to ORAC::Index in an attempt to
speed up ORAC::Index::indexentry.