scuba2: Attempt to make maps for DAISY scans with summit pipeline
The summit pipeline was using TCS_INDEX to decide when it could
make a map. This is fine for PONG maps where the index goes up
by one every rotation but for a DAISY observation TCS_INDEX is
one for the entire observation and only goes to 2 right at the
end as the timing dictates a slight overrun. This means that the
observer/tss have to wait up to 35 minutes to see what their
map looks like.

In the new scheme the pipeline determines how long it has been
since a daisy map has last been and if a threshold is exceeded
a map is made automatically.

Additionally, the pipeline now tries to detect when the observation
has ended (by comparing file lengths) and can now trigger a final
map. This needs some thought as there is a risk that a 10 second
map will be added into the coadd so we may want this to happen
if a minimum time since the last map has been exceeded.