lib: Remove the special case for warning filehandle
This reverts 4a3d0e20bd86217f943036e0e9baa6520c8de25e and I'm
not sure the rationale for that commit from ten years ago. I'm
sure we'll find out soon but the code has:

  if(!defined(@_)) { return; }

and defined(@_) doesn't really mean what I think is meant here.
I think it was meant to be "defined($_[0])" but I'm still not
sure why warhdl() gets to be different to errhdl() and outhdl().