SCUBA2: remove summit pipeline check
When the noise-processing code was first written, it was envisaged
that the summit pipeline would wait until all the noise data had been
taken before processing. However, there's no need to do that and, in
practice, would not have worked anyway because i) dark noises tend to
consist of a single file and ii) sky noises finish with a short dark
frame.