SCUBA2: establish a dynamic default fastflat type to return
Retrieving an appropriate dark fastflat for dark frames has been hard
to get right with the existing pipeline. Rather than workout what type
of raw fastflat to retrieve before calling this primitive, use
knowledge of the data to determine a suitable default value.

Normally, for on-sky / science data, the raw fastflat from earlier in
the same observation is required. For dark frames that are part of an
open-shutter observation, the only way to go is to retrieve a dark
fastflat from the most recent SETUP. In principle, it should retrieve
a dark fastflat from a more recent dark noise, but in practice dark
noise observations are only made towards the beginning and end of a
night, and are rarely inserted during the course of normal observing.