SCUBA2: restrict peak search for calibrators
There's no point searching the entire image to find the peak if we're
observing a calibrator. Try a small box around the map centre
instead. This should avoid those error messages in the pipeline where
it's trying to fit a bright edge pixel and coming up with a position
that's outside the map.

The situation for other targets is less clear, though there is an
argument for using the map size as defined in the header to try and
avoid noisy edge pixels.