SCUBA2: do not make a map if the carousel is closed
There is no information in the headers that explicitly states the
telescope carousel is closed. However, such observations are generally
made as part of the JCMTCAL and E&C projects so check for those
IDs. If we have such data then check the WVMTAU headers - the start
and end values will be identical as the WVM is not being updated.

It remains to be seen if this is a unique test.

Note that it is simplest to perform this check at the map-making stage
in order to allow other observation and sequence types (noise,
flatfield) to be processed as they are still potentially useful.