SCUBA2: write a dummy flag file if none exists at end of observation
The QL pipeline has an issue whereby the processing time for fastflats
and/or noise data increases to the point where crucial (usually
on-target) data are skipped. This can result in an incomplete focus
cube or no map being produced in the case of a pointing
observation. The pipeline only creates a flag file for the
POINTING_FOCUS task if a cube or image was made which led to the
situation of the PF task waiting for a flag file that was never going
to appear. In order to prevent this from happening, the pipeline
detects if the observation has ended without creating a pointing/focus
flag file. If so, then a dummy file is written to disk which the PF
task can read immediately.