SCUBA2: no need to trim the wcs just to determine the peak
A bug in the NDF library means that the "trim trimwcs" isn't working
properly at the moment, so remove those arguments from the call to
ndfcopy. However, it is really not necessary to reduce the number of
axes at this point just to find the peak in the map so it is unlikely
this change will need to be reinstated later.

Also protect against an undefined value returned for the statistics of
the focus positions.