SCUBA2: add option to calculate s/n for Group matched filtered image
After running the matched filter, the filtered image is stored as the
second file in the Grp. Allow the user to use that file with the
USE_MF parameter, only accessed if the current object is a Group.

Also return the hash reference with the statistics in the primitive
hash with the key RESULTS.