SCUBA2: add support for makemap REF parameter
If the user wants to use an external mask to constrain the map, then
the REF parameter must be defined and contain the name of that mask
file. Allow the REF to be given as a recipe parameter and always use
that if it is defined.
(cherry picked from commit 237fa9c689b3ee5e7aca9e6df5844d95ed336674)