SCUBA2: don't skip Frames with mixed sequence type
Previously this primitive only proceeded with looking for existing
solutions if the sequence type was defined in the header (ie not a
subheader). This meant that the summit pipeline wasn't always able to
look for an existing fastflat. On the other hand, this change probably
means that the offline pipeline performs an unnecessary search
(cherry picked from commit 74b7f42a3cf4288630ce24e2ac04e32fb5f08ede)