SCUBA2: perform QA analysis on flatfield results
To get the QA flatfield pass/fail result into index.flat, the QA must
be carried out at the same time as the rest of the analysis. The
slight downside is that the QA is carried out 4 times, but that's the
only way to do it.
(cherry picked from commit c72fc03580af7527414410ab701652d4be216e45)