heterodyne: Restructure to perform non-linearity tests on full observations.
Before the analysis was performed on each sub-file of an observation.
The new approach concatenates the non-linearity profiles.  This yields
more reliable base levels to discriminate non-linear baselines from
the noise-generated deviations, especially in shorter-duration
sub-files.  In the case of _MASK_BAD_BASELINE_LINEARITY_ it was
possible to have a sub-file with a section of non-linearity
interference exceeding the global threshold because it occupied
proportionally large part of the sub-file, and hence the whole
receptor would be removed from the reductions.

I did apply the same approach to the high-frequency noise filter, but
SMURF:FIXSTEPS refuses to process the concatenated and reshaped noise
profile.  Since the discrimation is clearcut and the contimated
spectra are in narrower time bands, it may matter little.