heterodyne: Perform the global non-linearity test after filtering transient noise.
Although now partly mitigated by the restructuring it is possible to
reject a receptor for transient interference depending on the degree
and relative durations.  This change removes those sections of time
series before judging the receptor as a whole, and improves the signal
to noise.  Reordering still catches the case where nothing is rejected
as transient interference because a receptor's spectra exhibit similar
non-linearity.