SCUBA2: use matched-filtered map if requested
Using the matched-filtered map is a little more complicated than just
setting a flag. First, the flag has to be set correctly ($use_mf was
incorrectly localized within a separate block when set by a primitive
argument); Second the correct file must be chosen if using the current
Group object.

Also append "_snr" suffix rather than replace it.  Since this file is
pushed onto the intermediates array and is therefore probably deleted
at the end of processing, it could be argued that this is
unnecessary. However, the PICARD default is now to append suffices,
and this primitive is used by PICARD too, so do that here.
(cherry picked from commit 4f11c086140b10094eebec35228be851860b64db)