SCUBA2: search in multiple locations for the given config file
This patch adds generalized support for directories that act as a
central store for the particular config file in question. The user
specifies a config type to the primitive, and the environment variable
defined by that config type with "_CONFIG_DIR"
(e.g. "FINDCLUMPS_CONFIG_DIR") is searched for the given config file.

The search order is: current working directory, ORAC_DATA_OUT and
dedicated config dir.