SCUBA2: store SNRMAP flag in Frame uhdr
If the creation of a S/N map was successful, and that image is being
stored in the current (Group or Frame) object, then set the SNRMAP
uhdr flag to 1. Then downstream primitives will be able to determine
if they are given a S/N image.
(cherry picked from commit 92c0e70c73317c553be35cdbec331e3c5afe52dd)