SCUBA2: assume RMS of SNR map is 1
Just assume that the RMS is unity if we are processing a S/N map,
rather than calculating it explicitly. This may not always be correct,
in that the distribution of image-plane values can have an RMS
different from 1, but in the minds of most astronomers it's probably
the right thing to do.
(cherry picked from commit 42f78851f680f14dbf3ef8f2406ee46b1ccf252d)