heterodyne: Offer new TILE recipe parameter and associated primitive arguments.
If TILE is set to 0, then tiling in MAKECUBE is disabled.  This was
requested by Jess, although I was already implementing it to compare
the spectral cubes of the JLS l=40.0 region to determine the cause of
the strange shuffling and wide-rangiong noise levels of data in the
outer eight tiles.  The l=40.0 cube looks normal when MAKECUBE does
not tile this region.
(cherry picked from commit 5d4b540964174782beacabfc5a3f577a65b1ec62)