lib: Create method for syncing product change to files
There was a bit of code that went

  - read allow_sync value
  - Allow sync
  - Update product
  - Update file
  - reset allow_sync

so there is now a new force_product_update() method to do this.
This method also handles the case where you have created a
dummy frame object to hide the file (files that looked like
raw files are not normally written to). This therefore
fixes a problem where the _snr product in the SCUBA-2 primitives
were not getting the "snr" product header updated.