heterodyne: Restore checking for sensitive FIXSTEPS.
FIXSTEPS sometimes seems a little fussy rejecting some of the profiles
generated by the Laplacian filter.  Looking at the profiles it's hard
to tell why, and for the cases observed so far ignoring the lack of
step correction has not proved to be an issue.

I thought this fix was already committed...  This probably got lost
while I was attempting to make a version that generates a single
profile across the subscans, but which did not work.