SCUBA2: do not use "veryshort" config file
Sometimes the final file in a sequence is short enough to convince the
pipeline that the "veryshort" config file should be used. In practice,
this is unlikely to be the case, so if this primitive detects a short
file, make sure that the "normal" config file type is used instead.