SCUBA2: do not calculate or store the mode
The distribution of bolometer noise, NEP and responsivity values can
occasionally trip up KAPPA histat. The error is trapped by the
pipeline, but it still generates red text and doesn't actually
indicate a problem with the data. While alternative methods of
estimating the modal value were examined, in the end it was decided
that the mode isn't actually used for anything (for example, the
median is used in QA) so its calculation can be removed altogether.