SCUBA2: trim redundant WCS axes before finding the peak
Downstream primitives for fitting the peak assume that the input image
is 2-d only - ie it uses a 2-d section. However, doing so for a 3-d
file means that stats returns blank results because the pixel
coordinate for the redundant axis is not 1. Therefore, since this
image is only used as for peak-finding, trim the redundant WCS axes
before analysis.

This eliminates the occasional warning seen in the pipeline where the
fitted peak exceeds the measured map peak by umpty-zillion-sigma (the
map peak was bad of course).