PICARD: read actual image dimensions for comparing with desired output size
Previously this primitive trusted the FITS header values for the map
size as being representative of the actual map size, which is close
enough to the truth for large PONG maps, but is not the case for DAISY
scans. Use the actual image dimensions instead to make the assessment.

Update POD to describe a potential gotcha for small crops.