SCUBA2: use new input arguments to control processing
Now this primitive uses the files in the current frame object with the
option of specifying a SMURF extension (via the new SMURF argument),
e.g. "nep" to mosaic the NEP data. The output mosaic has a suffix
given by the new SUFFIX parameter and is stored in the Frame at the
end of processing. The file is tagged with the given suffix.

The INFILES and OUT parameters have been removed.