SCUBA2: tag and store matched-filtered output as a named product
The product is defined as "Matched-filtered" and is used to check the
input data so that the filter is not applied twice. The output file is
tagged (using the simpler name "matchfilt") so it can be used by
subsequent primitives.

The OUT, OBSEND and STORE arguments have been removed - the filtered
input map is now stored in the current Frame or Group object (as
defined by the GROUP parameter). Similarly, the inout method for the
current object is used to define the name of the filtered map.

A check has been added for the case when the user supplies a PSF that
the pixel scale matches that of the input file.