SCUBA2: input Frm/Grp must now contain the data and model
The IMAGE and MODEL input parameters have been removed so that the
input Frame or Group (determined by the GROUP parameter) must contain
two files: the data and the model fit. These files are retrieved using
the tag "profile_fit".

A check is made to see if the current product for the input object
indicates matched-filtered data so that corresponding plot titles can
be set and the data are plotted correctly.