SCUBA2: calculate SNR image for whatever is in given object
Now this primitive simply calculates an SNR image for the file in the
specified object. If the user wishes that to be the output from the
matched filter then they must set the input file accordingly.

Output data are tagged with "snr" (same as product name) and is always
updated.

The SNRMAP uhdr has been removed as the tag serves the same purpose
but in a better way.