SCUBA2: use tags for retrieving/storing data
Rather than mess about with uhdr flags, tags are created for each
focus position, all of the images to go into the cube, the midplane
image (used for determining the source position before shifting the
WCS) and the final output cube. The cube is also stored as the product
"focus_cube". Another feature of this is the replacement of temporary
files with named intermediates to aid with debugging.

On exit, the Frame object is restored to its entry state (the focus
cube is a Group product).

One last update is that the sky mapping is now only retrieved once.