PICARD: make sure all files to be stacked have same bounds
Stackframes requires that all the input files have the same bounds
which will be true for focal-plane mosaics, but not for maps of
astronomical sources. Use wcsmosaic to determine the encompassing
bounds and use that as a template for all the files going into the
stack. Note that the wavelength axis is removed to ensure the output
has only 3 dimensions.

Also the default behaviour is to stack all images, and not separate
them by UTDATE. Use the recipe parameter STACK_KEYS instead.