PICARD: use newimage and snr tags/products for input files
The input files are tagged with "newimage" (replacing the NEWIMAGE
uhdr entry), and the Frame product is set to the same.

The units of the input file are checked and if blank, it is assumed
that the input data are SNR images and the data are tagged with "snr"
(and product set the same).