SCUBA2: check for badpeak tag before continuing
If the tag "badpeak" exists for the current Frame then the file has
bad values at the peak. In this case unset all of the process control
flags and let the primitive run as a no-op. The values stored in the
primitive hash will therefore be undefined so calling primitives
should check before attempting to use them (but they should be doing
this anyway).