SCUBA2: use a copy of the group file in the coadd list
When coadding in running mode, the current group file is included in
the list of files to coadd. This can mean that the file ends up in the
intermediates array where it may be deleted by mistake. This may be
prevented by making a copy of the group file before coadding.