docs: add Makefile to build documentation.
There are 3 main rules: clean, build and install.  Since the
LaTeX files derived from code do not (yet) have dependencies
specified, running 'make clean' is recommended before building.

The Makefile performs the steps from a shell script used
for the Kapuahi release, which are described on the StarlinkRelease
wiki page, with the addition of .htx.build files to track
the building of the HTML documentation.