SCUBA2: deal with case where there are no good bolometers from setup
If the setup fails, the flatfield ramps will return zero good
bolometers. In this case we can still write out the names of the
flatfield files to a flag file, but issue an error stating that it
will be incomplete.

Furthermore, it is not possible to write or update a badbol file if
there are no good bolometers. Let the existing list be used instead.