SCUBA2: add option to trim image to default map size
Running a source finder on an image runs the risk of detecting false
sources near the edge of the mapped area. In this case it may be more
appropriate to trim the map first, especially in cases where the
number of sources is small and expected to lie somewhere near the map
centre.

Add the CROP parameter which indicates whether the current map should
be trimmed before running the source finder. The default is false. If
true, then the pipeline uses the default map size as given in the
header.