SCUBA2: also pass flatfields to sc2concat.
The _CONCAT_FTS_SCANS_ primitive uses the 'flat=true' option
of sc2concat to ensure that the data are flatfielded.  However,
sc2concat was not finding the correct flatfields to apply.  This
patch collects fastflats by subarray and observation and passes
them all explicitly to sc2concat along with the data files.