SCUBA2: restore current Frame to original state before proceeding
It was noticed that the data points derived from the final two SMU
positions were identical. This was tracked down to the fact that the
subgroup for the final SMU position cointained Frm, which had been
used multiple times to store the files at each SMU position. When it
came to the final SMU position, Frm now contained the file stored by
the previous position rather than the file for the current SMU
position. This is easily solved by restore Frm to its original state
before processing the next SMU position with a tagretrieve call.