SCUBA2: continue with FCF calculation even if source not detected
The FCF calculations only proceed if the source is detected (via a
gaussian fit) in the signal map. However, it is possible that the
source is still detectable via aperture photometry (albeit at low
signal-to-noise) and after the application of a matched
filter. Therefore, the pipeline should proceed with estimating those
FCFs even if the source could not be detected in the signal map.

Furthermore, the quoted FCFs are colour coded based on how close they
are to the standard FCF. If the values are within 5/10% (at 850/450
respectively), the values are green; yellow if within twice that spec,
and red if beyond that.