SCUBA2: change format of fcf log file, make it optional
The FCF log file now contains the FCFs and their respective errors on
the same line, and the derived effective FWHM if it could be
calculated. Undefined values are stored as NaN.

A new parameter called LOGFILE has been added which is a flag to
indicate whether or not the FCF log file should be written. Default is
to write the file, which is required by the pipeline.