PICARD: mark Frame as good to proceed with rest of recipe
If no FCFs were derived, the Frame is marked as bad by the
pipeline. This excludes the data from some further analysis which uses
SCUBA2 primitives. This is fine for the SCUBA2 pipeline which may have
good reasons for excluding those data from further processing, but in
PICARD-land we want to continue with our analysis.