SCUBA2: update user message when running makemap
Previously the user was warned that things may take a very long time
if the number of files exceeded 16. The trouble is that then the user
gets that message if they're processing 17 files or 280, which isn't
very helpful. Now the message is displayed if the number of files
exceeds 64, and emphasized if more than 128.