PICARD: rename to better reflect what this recipe should be used for
The old FCFNEFD recipe was cobbled together during commissioning and
was never a coherent piece of work. With the upcoming Starlink release
and data redeuction workshop, a new improved version was required that
could be used by observers to investigate the state of the calibration
observations obtained on the same night as their data. Behold,
SCUBA2_CHECK_CAL - perform checks on calibrator data.