Calib::SCUBA2: new methods for handling fwhm and error beam
The results of a beam calculation are stored in the calibration object
in the hash reference BeamPar. Unfortunately, retrieving those beam
parameters requires unpacking the hash reference by the user, so two
new methods have been added to provide direct access to the latest
values of the FWHM and the error beam contribution.

Both of these are now calculated automatically if the beampar method
returns a valid result, and are updated at the time new beam
parameters are stored.