SCUBA2: update when to invert ZPD interferograms.
Used the criteria from Brad Gom's email of 2 Jan 2013 to
determine when to invert the ZPD interferogram in
_FIT_INTERFEROGRAM_PEAK_:

With the S8D shutter in, the S8D output port is dominant
(i.e. fringe is positive).
With the S8C shutter in, the S8C output port is dominant
(i.e. fringe is positive).

The shutter situation will be easier to determine once the
corresponding FITS header is added, but we may need to do
it this way for "old" data anyway.