SCUBA2: shortened names for the faint point sources recipes
It's one thing having long recipe names that are automatically
selected, but expecting users to type them in is perhaps a bit
much. Therefore create shortened versions that still describe the
essence of what the recipe is about. The exception is REDUCE_CLS of
course which was previously renamed and for now, it's exactly the same
recipe.