SCUBA2: add ability to be used by offline pipeline
Previously this primitive was only used by the QL pipeline to
determine the noise properties of science data. Now it can be used by
the offline pipeline (with a suitable recipe), and can either operate
on the first on-sky subscan or the entire timeseries.

Also fix a bug whereby the noise data weren't actually being used in
the QA by delaying the QA reporting for when the noise data are being
analyzed. This works because the weighted NEP determined in the first
call to _ANALYZE_NOISE_ is still stored in the QA report info hash.