SCUBA2: new recipe to allow offline noise analysis of science data
This recipe more-or-less repeats the analysis performed by the QL
pipeline, enabling observers to examine the noise properties of their
data and get a handle on the data quality.

The users can specify a config file to use in the noise calculation,
as well as the frequency range to be used. By default, the recipe will
only use the first 30 seconds of on-sky data, but the user can force
the entire observation to be used instead.