SCUBA2: write log file + store noise mosaics for later use
This primitive now writes out a customized log file summarizing the
results of the QA calculations in a way that allows easy plotting with
Topcat. The log file is named according to the current wavelength so
450 and 840 data are logged separately.

The noise mosaics are now stored and tagged with "scinoi" so they can
be retrieved at a later date and turned into a 3-d cube.