SCUBA2: check for minimum default map size, set to 60" if too small
Some maps have MAP_HGHT and MAP_WDTH FITS headers that are too small
to allow sensible processing and analysis. Therefore in the cases where
there is no new width or height specified, make sure that the default
values are not too small. For now, force a minimum ma size of 60".