SCUBA2: use linear regression in APPROX_FTS_FRAME
This primitive now performs linear regression between the
pixel coordinates and the FTS moving mirror positions
near to the estimated ZPD position.  This should give a more
accurate approximation of the position and avoids the
problems of using fixed offsets.  The region over which
the fit is performed is that between the two baseline
regions, i.e. the extimated ZPD position +/- the
FTS_ZPD_BASE_OFFSET parameter.