SCUBA2: scale the default box size by the actual pixel size
The assessment of whether a source is point-like is done in pixel
space. An image of a point source made with 1" pixels at 850 um does
not pass the point-source test on account of the source covering many
more pixels than assumed. Thus the default box size should take into
account the pixel scale such that maps with small pixels use a larger
default box size.

This scheme should still work for large pixels, though it has not been
tested.

It is still assumed that a user-provided value for box is correct.