SCUBA2: increase the default map size to 75 arcsec
Due to a strange interaction between this default map size and the
default search area for _CENTROID_IMAGE_, certain maps fail the
criteria for proceeding with a centroid estimate, even if the data are
perfectly good. The simplest solution is to increase the default map
size to something not so easily related to the default centroid box.