SCUBA2: use a pixel-scale dependent box size for fitting
A box of 9 pixels doesn't always work for maps made with small pixels,
and even increasing it slightly does not ensure that all fits
succeed. So implement a pixel-scale dependent value so that the same
spatial area is fitted.

(I'm beginning to think that the pixel scale should be determined and
stored in the header (uhdr). I don't know of a case where the pixel
scale changes during processing, so this should be OK to implement.)