SCUBA2: always re-read header after storing noise results
Previously the pirmitive only re-read the header if the number of
files stored was >1. However, that statement assumed the case of n=1
corresponded to there only being 1 input file, and not that 3 of 4
subarrays had failed to produce results.

Since re-reading the header for a single file carries little overhead,
it seems worth re-reading the header in all cases so avoid potential
downstream "files_from_hdr" errors,