PICARD: add support for two-component PSFs
New parameter: PSF_RELAMPS - a comma-separated pair of values
indicating the relative amplitudes of the gaussians specified by
PSF_FWHM.

PSF_FWHM can now be specified as a comma-separated pair of values.

Note that if two values are given for PSF_FWHM but there is no
corresponding PSF_RELAMPS, then the standard wavelength-appropriate
values for the telescope beam will be used.