SCUBA2: add support for two-component beam
This commit contains major updates to the SCUBA2 calibration class,
including the renaming of several methods.

Whereas previously the class did not differentiate between the actual
telescope beam and a fit to the beam, methods associated with fits
have "fit" in their names.

The default two-component telescope beam properties are stored in a
hash indexed by wavelength (actually the value returned by the
subsint() method, as used by the default FCFs). The keys are FWHM1,
FWHM2, AMP1 and AMP2. The amplitudes are the relative fraction and add
to unity.

Details follow:

 * beam - now returns the hash reference containing the telescope beam
   parameters.

 * beamamps - new method to return just the relative amplitudes of the
   telescope beam.

 * beamfit (renamed from beampar) - stores/returns the fitted beam
   parameters. Note this method only stores the results for a single
   component fit.

 * fwhm - now returns an array (or array ref) with the FWHM for each
   telescope beam component.

 * fwhm_eff (renamed from fwhm) - returns the FWHM of a gaussian with
   the same area as the telescope effective beam area.

 * fwhm_fit - return the FWHM derived from the most recent fit.

 * store_beam - has been removed.

Finally, the methods have been re-grouped and reordered alphabetically.