SCUBA2: add ability to create a two-component PSF
The inputs to this primitive now permit two values to be given for the
FWHM parameter, along with corresponding relative amplitudes specified
by the RELAMPS parameter. Both are overridden by the preseence of
their respective recipe parameters.

Note that a single value for the FWHM is still accepted.

The default behaviour is now to create and use a two-component PSF.