heterodyne: Add minimum spectra argument (MIN_SPECTRA).
Small numbers of spectra either cause certain tasks like FINDBACK or
FIXSTEPS to abort with an error, and there needs to be sufficient
spectra to define a norm for non-linearity above which anomalous
spectra can be rejected.  For now set the default to 50 spectra, and
see how it works in practice.  There is a case to make the threshold
somewhat larger.