SCUBA2: set parameters for fake gaussian if none given
It is assumed that if the user is calling a recipe where a fake source
is to be added to the timeseries then this primitive should always add
one, whether or not the user has specified parameters.

In the case where no FAKEMAP recipe parameters are given, the
primitive creates a fake point source using the FWHM at the current
wavelength and an amplitude of 10/50 Jy/beam at 850/450 um
respectively.