SCUBA2: factor out some steps to other primitives
The basic map-making step is now not part of this primitive, so it
should be included in the corresponding recipe. There is an issue in
making sure that both primitives use the same config file but I'll
address that later.

The code setting the properties of the fake source has been moved in to
_FAKEMAP_STEER_.

The MAP_PSF and KEEP_SN input arguments have been removed. I can't
imagine a reason this primitive would be called if the user did not
want a map psf. KEEP_SN is effectively handled in the steering
primitive.

The Grp uhdr entry FAKESCALE is set so that subsequent calls will use
the same value for scaling the artificial gaussian.