SCUBA2: do not store files in current Frame
This commit largely reverts b4323dfa15 as I discovered an issue in
PICARD whereby the input list of files contained the original input
files, which resulted in those files being added to the intermediates
list of the current Frame and were therefore being deleted at the end
of processing.

However, all is not lost - the list of files in the current Frame is
still used as the default input list.