PICARD: push files onto Frm intermediates for quieter deleting
When calling _DELETE_TEMP_FILES_ with VERBOSE=0 and ALSOGRP=1, messages
about deleting Frame-based intermediates are suppressed, while those
pertaining to Group-based intermediates are not.

There can be a lot of aligned files from this primitive which should
be deleted, but it's distracting and unecessary to have the recipe end
with 100+ messages about deleting intermediate files which the user
didn't know existed and doesn't care about.