PICARD: new recipe/primitive to produce an effective PSF from sutiable files
This new recipe allows the user to take the _mappsf files generated by
REDUCE_SCAN_FAINT_POINT_SOURES_JACKKNIFE and create a correctly-scaled
coadded effective PSF that can be passed to SCUBA2_JACKKNIFE for
investigating the effects of varying various jack-knife parameters.

The scaling factor for the PSF will be obtained from the FITS header
of the input files if possible, otherwise the recipe parameter
FAKEMAP_SCALE must be specified.