SCUBA2: experimental feature to trim the noisy edges from the coadded map
If the recipe parameter THRESH is given, then the noisy edges are
trimmed from the map. The procedure determines the median exposure
time in the coadd and removes all pixels with an exposure time less
than THRESH times the median. A good starting place for THRESH appears
to be 0.5. Further experimentation will be needed to decide the
robustness of this estimator.

Alternative estimators may also be given, if desired, via the
STATS_ESTIMATOR recipe parameter, which may be mean, median or mode.